@@ -2,40 +2,38 @@ import 'package:flutter/cupertino.dart';
 
 class TagContainer extends StatefulWidget {
   final String tag;
-  final EdgeInsets tagContainerPadding;
-  final EdgeInsets tagContainerMargin;
-  final TextStyle tagTextStyle;
+  final EdgeInsets? tagContainerPadding;
+  final EdgeInsets? tagContainerMargin;
+  final TextStyle? tagTextStyle;
 
-  final int tagTextMaxLines;
-  final TextOverflow tagTextOverflow;
-  final TextAlign tagTextAlign;
-  final bool tagTextSoftWrap;
-  final Locale tagTextLocale;
+  final int? tagTextMaxLines;
+  final TextOverflow? tagTextOverflow;
+  final TextAlign? tagTextAlign;
+  final bool? tagTextSoftWrap;
+  final Locale? tagTextLocale;
 
-  final GestureTapCallback onPressed;
-  final GestureTapCallback onLongPressed;
-  final GestureTapCallback onDoubleTap;
-  final Icon tagIcon;
-  final BoxDecoration tagContainerDecoration;
+  final GestureTapCallback? onPressed;
+  final GestureTapCallback? onLongPressed;
+  final GestureTapCallback? onDoubleTap;
+  final Icon? tagIcon;
+  final BoxDecoration? tagContainerDecoration;
 
   TagContainer(
-      {Key key,
-        @required this.tag,
-        this.onPressed,
-        this.onDoubleTap,
-        this.onLongPressed,
-        this.tagIcon,
-        this.tagContainerMargin,
-        this.tagContainerPadding,
-        this.tagTextStyle,
-
-        this.tagTextMaxLines,
-        this.tagTextOverflow,
-        this.tagTextAlign,
-        this.tagTextLocale,
-        this.tagTextSoftWrap,
-
-        this.tagContainerDecoration})
+      {Key? key,
+      required this.tag,
+      this.onPressed,
+      this.onDoubleTap,
+      this.onLongPressed,
+      this.tagIcon,
+      this.tagContainerMargin,
+      this.tagContainerPadding,
+      this.tagTextStyle,
+      this.tagTextMaxLines,
+      this.tagTextOverflow,
+      this.tagTextAlign,
+      this.tagTextLocale,
+      this.tagTextSoftWrap,
+      this.tagContainerDecoration})
       : super(key: key);
 
   @override
@@ -44,8 +42,8 @@ class TagContainer extends StatefulWidget {
 
 class _TagContainerState extends State<TagContainer>
     with TickerProviderStateMixin {
-  AnimationController _animationController;
-  Animation<double> _animation;
+  late AnimationController _animationController;
+  late Animation<double> _animation;
   static const int _fadeDuration = 450;
 
   @override
@@ -82,8 +80,16 @@ class _TagContainerState extends State<TagContainer>
             mainAxisAlignment: MainAxisAlignment.center,
             mainAxisSize: MainAxisSize.min,
             children: [
-              Text(widget.tag, style: widget.tagTextStyle, maxLines: widget.tagTextMaxLines, overflow: widget.tagTextOverflow, textAlign: widget.tagTextAlign, softWrap: widget.tagTextSoftWrap, locale: widget.tagTextLocale,),
-              widget.tagIcon != null ? widget.tagIcon : SizedBox()
+              Text(
+                widget.tag,
+                style: widget.tagTextStyle,
+                maxLines: widget.tagTextMaxLines,
+                overflow: widget.tagTextOverflow,
+                textAlign: widget.tagTextAlign,
+                softWrap: widget.tagTextSoftWrap,
+                locale: widget.tagTextLocale,
+              ),
+              widget.tagIcon != null ? widget.tagIcon! : SizedBox()
             ],
           ),
         ),
